[WK2] Regression(r131990): plugins/npruntime/remove-property.html started failing https://bugs.webkit.org/show_bug.cgi?id=99977
Reviewed by Alexey Proskuryakov.

Source/WebKit2:

Remove int32_t casts - the ArgumentDecoder::decode overload that takes a const int would cause
a temporary to be created which the number was then decoded into.

* Shared/Plugins/NPIdentifierData.cpp:
(WebKit::NPIdentifierData::encode):
(WebKit::NPIdentifierData::decode):

LayoutTests:

Remove now passing test.

* platform/efl-wk2/TestExpectations:

Canonical link: https://commits.webkit.org/118040@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@132120 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebKit2/Shared/Plugins/NPIdentifierData.cpp
@@ -73,7 +73,7 @@ void NPIdentifierData::encode(CoreIPC::ArgumentEncoder* encoder) const
     if (m_isString)
         encoder->encode(m_string);
     else
-        encoder->encode(static_cast<int32_t>(m_number));
+        encoder->encode(m_number);
 }
 
 bool NPIdentifierData::decode(CoreIPC::ArgumentDecoder* decoder, NPIdentifierData& result)
@@ -84,7 +84,7 @@ bool NPIdentifierData::decode(CoreIPC::ArgumentDecoder* decoder, NPIdentifierDat
     if (result.m_isString)
         return decoder->decode(result.m_string);
 
-    return decoder->decode(static_cast<int32_t>(result.m_number));
+    return decoder->decode(result.m_number);
 }
 
 } // namespace WebKit
